+++ R/app_server.R
@@ -778,10 +778,10 @@ app_server <- function(input, output, session) {
         return()
       }
       
-      ShowPreselectedSigs(input, output,input.catalog.type())
+      selected.sig.universe <- ShowPreselectedSigs(input, output,input.catalog.type())
       # Each time when we call ShowPreselectedSigs(), must remember to update
       # the reactive value preselected.sigs
-      preselected.sigs(input$preselectedSigs)
+      preselected.sigs(selected.sig.universe)
     
       # Show the actionButton for user to add more signatures
       output$addSig <- renderUI(
@@ -893,10 +893,10 @@ app_server <- function(input, output, session) {
         catalog <<- list.of.catalogs[[catalog.name]]
       }
       
-      ShowPreselectedSigs(input, output, input.catalog.type())
+      selected.sig.universe <- ShowPreselectedSigs(input, output, input.catalog.type())
       # Each time when we call ShowPreselectedSigs(), must remember to update
       # the reactive value preselected.sigs
-      preselected.sigs(input$preselectedSigs)
+      preselected.sigs(selected.sig.universe)
       
       # Show the actionButton for user to add more signatures
       output$addSig <- renderUI(
@@ -917,10 +917,10 @@ app_server <- function(input, output, session) {
         catalog.name <- paste0("cat", input.catalog.type())
         catalog <<- list.of.catalogs[[catalog.name]]
       }
-      ShowPreselectedSigs(input, output, input.catalog.type())
+      selected.sig.universe <- ShowPreselectedSigs(input, output, input.catalog.type())
       # Each time when we call ShowPreselectedSigs(), must remember to update
       # the reactive value preselected.sigs
-      preselected.sigs(input$preselectedSigs)
+      preselected.sigs(selected.sig.universe)
       
       # Show the actionButton for user to add more signatures
       output$addSig <- renderUI(