+++ NEWS.md
@@ -11,6 +11,9 @@ catalog type.
 
 * Fixed a bug in function `ReadAndCheckCatalog` for checking number of files to be read in.
 
+## Updated
+* Updated the height for SBS192 plot in `renderPlot`.
+
 <br>
 
 # mSigAct.server 0.0.8